changed uint96 to uint104 for account.frens

+++ contracts/facets/StakingFacet.sol
@@ -91,7 +91,7 @@ contract StakingFacet {
             s.tickets[id].accountBalances[msg.sender] += 1;
             s.tickets[id].totalSupply += 1;
         }
-        s.accounts[msg.sender].frens = uint96(frensBal);
+        s.accounts[msg.sender].frens = uint104(frensBal);
         emit TransferBatch(address(this), address(0), msg.sender, _ids, values);
         uint256 size;
         address to = msg.sender;
